add socket.io + connect mqtt
